requirements.txt
git+https://github.com/s4w3d0ff/python-poloniex.git
pyyaml
boto3
